refactor(modules): change modules structure
