--- brain_games/scripts/even.py
@@ -1,5 +1,0 @@
-def even(num):
-    if num % 2:
-        return "no"
-    else:
-        return "yes"
